requirements.txt
aiofiles==23.2.1
anyio==4.2.0
argcomplete==3.2.1
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
boto3==1.34.11
botocore==1.34.11
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dacite==1.7.0
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.2.0
fiftyone==0.23.2
fiftyone-brain==0.14.2
fiftyone-db==1.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
glob2==0.7
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
humanize==4.9.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
imutils==0.5.4
inflate64==1.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.4
mongoengine==0.24.2
motor==3.3.2
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
paho-mqtt==1.6.1
pandas==2.0.3
pillow==10.2.0
plotly==5.18.0
pprintpp==0.4.0
priority==2.0.0
psutil==5.9.7
py-cpuinfo==9.0.0
py7zr==0.20.8
pybcj==1.0.2
pycryptodomex==3.19.1
pymongo==4.6.1
pyparsing==3.1.1
pyppmd==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzstd==0.15.9
rarfile==4.1
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
s3transfer==0.10.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.34.0
strawberry-graphql==0.138.1
sympy==1.12
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==8.2.3
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.0.232
universal-analytics-python3==1.1.1
urllib3==1.26.18
voxel51-eta==0.12.2
wcwidth==0.2.12
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.17.0
